Updated template versions and fixed code
