+++ .github/workflows/blui-ci.yml
@@ -417,6 +417,8 @@ jobs:
               with:
                   node-version: ${{ matrix.node-version }}
                   registry-url: 'https://registry.npmjs.org'
+            - name: Install pnpm
+              run: npm install -g pnpm
             - run: pnpm install
             - name: Publish package
               run: npm run publish:package -- -b ${{env.BRANCH}}
