+++ packages/cli-templates/templates/authentication-typescript/package.json
@@ -14,27 +14,27 @@
     },
     "eslintConfig": {},
     "dependencies": {
-        "@brightlayer-ui/colors": "^3.0.1",
-        "@brightlayer-ui/icons-mui": "^3.6.0",
-        "@brightlayer-ui/react-auth-workflow": "^6.0.0",
-        "@brightlayer-ui/react-components": "^7.0.0",
-        "@brightlayer-ui/react-themes": "8.0.0",
+        "@brightlayer-ui/colors": "^3.1.1",
+        "@brightlayer-ui/icons-mui": "^4.0.0",
+        "@brightlayer-ui/react-auth-workflow": "^7.0.0-alpha.6",
+        "@brightlayer-ui/react-components": "^8.0.0-alpha.2",
+        "@brightlayer-ui/react-themes": "^8.0.0-alpha.1",
         "@emotion/react": "^11.7.1",
         "@emotion/styled": "^11.10.8",
         "@mui/base": "^5.0.0-beta.63",
-        "@mui/icons-material": "^6.1.9",
-        "@mui/material": "^6.1.4",
+        "@mui/icons-material": "^7.2.0",
+        "@mui/material": "^7.2.0",
         "@okta/okta-auth-js": "^7.10.1",
         "@okta/okta-react": "^6.9.0",
-        "@types/react": "19.1.8",
-        "@types/react-dom": "^18.0.0",
+        "@types/react": "^19.1.8",
+        "@types/react-dom": "^19.1.6",
         "@vitejs/plugin-react": "^4.3.1",
-        "date-fns": "^3.6.0",
-        "dotenv": "^16.4.5",
-        "i18next": "^23.7.13",
-        "react": "^18.2.0",
-        "react-dom": "^18.0.0",
-        "react-i18next": "^13.0.3",
+        "date-fns": "^4.1.0",
+        "dotenv": "^17.2.0",
+        "i18next": "^25.3.2",
+        "react": "^19.1.0",
+        "react-dom": "^19.1.0",
+        "react-i18next": "^15.6.0",
         "react-router-dom": "^6.8.0",
         "vite-tsconfig-paths": "^5.1.4"
     },
@@ -44,13 +44,13 @@
         "@emotion/css": "^11.11.2",
         "@testing-library/dom": "^10.4.0",
         "@testing-library/jest-dom": "^6.4.5",
-        "@testing-library/react": "^13.3.0",
+        "@testing-library/react": "^16.3.0",
         "@testing-library/user-event": "^14.6.1",
-        "@types/node": "^22.10.3",
-        "@typescript-eslint/eslint-plugin": "^5.62.0",
-        "@typescript-eslint/parser": "^5.49.0",
+        "@types/node": "^24.0.14",
+        "@typescript-eslint/eslint-plugin": "^8.37.0",
+        "@typescript-eslint/parser": "^8.37.0",
         "eslint": "^9.15.0",
-        "eslint-config-prettier": "^9.1.0",
+        "eslint-config-prettier": "^10.1.0",
         "eslint-plugin-prettier": "5.2.1",
         "eslint-plugin-react": "^7.37.4",
         "identity-obj-proxy": "^3.0.0",
@@ -58,7 +58,7 @@
         "prettier": "^3.3.3",
         "ts-node": "^10.9.2",
         "typescript": "^5.7.2",
-        "vite": "^5.4.19",
+        "vite": "^7.0.5",
         "vite-plugin-dts": "^4.3.0",
         "vitest": "^3.1.1"
     },
