+++ packages/cli-templates/templates/blank-typescript/package.json
@@ -3,17 +3,17 @@
     "version": "1.0.0",
     "private": true,
     "dependencies": {
-        "@brightlayer-ui/colors": "^3.1.1",
-        "@brightlayer-ui/icons-mui": "^3.6.0",
-        "@brightlayer-ui/react-components": "^7.0.0",
-        "@brightlayer-ui/react-themes": "8.0.0",
+        "@brightlayer-ui/colors": "^4.0.0",
+        "@brightlayer-ui/icons-mui": "^4.0.0",
+        "@brightlayer-ui/react-components": "^8.0.0-alpha.2",
+        "@brightlayer-ui/react-themes": "^8.0.0-alpha.1",
         "@emotion/css": "^11.13.0",
         "@emotion/react": "^11.13.3",
         "@emotion/styled": "^11.13.0",
-        "@mui/icons-material": "^6.1.4",
-        "@mui/material": "^6.1.4",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
+        "@mui/icons-material": "^7.2.0",
+        "@mui/material": "^7.2.0",
+        "react": "^19.1.0",
+        "react-dom": "^19.1.0",
         "vite-tsconfig-paths": "^5.0.1"
     },
     "scripts": {
@@ -37,10 +37,10 @@
         "@testing-library/react": "^16.0.1",
         "@testing-library/user-event": "^14.5.2",
         "@types/node": "^22.13.0",
-        "@types/react": "^18.3.4",
-        "@types/react-dom": "^18.3.0",
-        "@typescript-eslint/eslint-plugin": "^5.62.0",
-        "@typescript-eslint/parser": "^5.62.0",
+        "@types/react": "^19.1.8",
+        "@types/react-dom": "^19.1.6",
+        "@typescript-eslint/eslint-plugin": "^8.37.0",
+        "@typescript-eslint/parser": "^8.37.0",
         "@vitejs/plugin-react": "^4.3.3",
         "eslint": "^9.15.0",
         "eslint-config-prettier": "^10.1.1",
@@ -50,9 +50,9 @@
         "prettier": "^3.3.3",
         "ts-node": "^10.9.2",
         "typescript": "^5.7.2",
-        "vite": "^5.4.19",
+        "vite": "^7.0.5",
         "vite-plugin-dts": "^4.5.3",
-        "vitest": "^2.1.9"
+        "vitest": "^3.2.4"
     },
     "prettier": "@brightlayer-ui/prettier-config"
 }
